# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GHH Library Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/libraries:
    get:
      operationId: getAdminLibraries
      tags: [admin]
      summary: List all libraries
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Libraries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminLibrary"
  /admin/libraries/{id}:
    get:
      operationId: getAdminLibraryDetail
      tags: [admin]
      summary: Library detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Library detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLibrary"
        "404":
          description: Not found
    patch:
      operationId: updateAdminLibrary
      tags: [admin]
      summary: Update library status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLibrary"
  /admin/students:
    get:
      operationId: getAdminStudents
      tags: [admin]
      summary: List all students across all libraries
      parameters:
        - name: library
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminStudent"
...
  /admin/attendance:
    get:
      operationId: getAdminAttendance
      tags: [admin]
      summary: Attendance logs
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: library
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Attendance logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceLog"
  /admin/notifications:
    get:
      operationId: getAdminNotifications
      tags: [admin]
      summary: Notifications list
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: sendAdminNotification
      tags: [admin]
      summary: Send a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ── Library Owner Dashboard ────────────────────────────────────────
...
components:
  schemas:
...
    AdminLibrary:
      type: object
      required:
        - id
        - name
        - ownerName
        - mobile
        - city
        - totalSeats
        - activeStudents
        - status
        - registrationDate
        - monthlyRevenue
      properties:
        id:
          type: string
        name:
          type: string
        ownerName:
          type: string
        mobile:
          type: string
        city:
          type: string
        area:
          type: string
        totalSeats:
          type: integer
        activeStudents:
          type: integer
        status:
          type: string
          enum: [approved, pending, rejected, suspended]
        registrationDate:
          type: string
        monthlyRevenue:
          type: number
        occupancyRate:
          type: number
        facilities:
          type: array
          items:
            type: string
    LibraryStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, pending, rejected, suspended]
    AdminStudent:
      type: object
      required:
        - id
        - name
        - email
        - mobile
        - library
        - seat
        - plan
        - creditsRemaining
        - attendancePercent
        - status
        - joinDate
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        library:
          type: string
        seat:
          type: string
        plan:
          type: string
        creditsRemaining:
          type: integer
        attendancePercent:
          type: number
        status:
          type: string
          enum: [active, suspended, expired]
        joinDate:
          type: string
...
    AttendanceLog:
      type: object
      required:
        - id
        - studentName
        - library
        - date
        - entryTime
        - shift
        - status
      properties:
        id:
          type: string
        studentName:
          type: string
        library:
          type: string
        date:
          type: string
        entryTime:
          type: string
        exitTime:
          type: ["string", "null"]
        shift:
          type: string
        status:
          type: string
          enum: [present, absent, leave, late]
        creditsDeducted:
          type: ["number", "null"]
    Notification:
      type: object
      required: [id, title, message, target, sentAt, type]
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        target:
          type: string
          enum: [all, library, student]
        targetId:
          type: ["string", "null"]
        sentAt:
          type: string
        type:
          type: string
          enum: [announcement, alert, reminder, promotion]
    NotificationInput:
      type: object
      required: [title, message, target, type]
      properties:
        title:
          type: string
        message:
          type: string
        target:
          type: string
          enum: [all, library, student]
        targetId:
          type: ["string", "null"]
        type:
          type: string
          enum: [announcement, alert, reminder, promotion]

    # ── Library Owner ────────────────────────────────────────────────
...