# lib/api-zod/src/generated/types/adminLibrary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GHH Library Management API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminLibraryStatus } from './adminLibraryStatus';

export interface AdminLibrary {
  id: string;
  name: string;
  ownerName: string;
  mobile: string;
  city: string;
  area?: string;
  totalSeats: number;
  activeStudents: number;
  status: AdminLibraryStatus;
  registrationDate: string;
  monthlyRevenue: number;
  occupancyRate?: number;
  facilities?: string[];
}