# lib/api-zod/src/generated/types/adminStudent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GHH Library Management API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStudentStatus } from './adminStudentStatus';

export interface AdminStudent {
  id: string;
  name: string;
  email: string;
  mobile: string;
  library: string;
  seat: string;
  plan: string;
  creditsRemaining: number;
  attendancePercent: number;
  status: AdminStudentStatus;
  joinDate: string;
}